Use same directories in nautilus bar and app
We were using different ways to get the public and downloads dir.
Whereas the main user-share had fallbacks to avoid sharing $HOME,
the nautilus bar was using the special dirs without any fallbacks.

This moves the common code to user_share-common.[ch] and makes sure
that fallbacks are used in both cases.