Fix http sharing not workin on Linux 3.0
gnome-user-share tries to discover a free port to run Apache on by
bind()ing a struct sockaddr_in with no port specified in
src/http.c:get_port(). However, it neglects to specify an address
family, so that attempt fails with EINVAL, and the -1 returned gets
passed to spawn_httpd() as the port that Apache should be listening on,
which causes to Apache promptly and silently bail.

This error was caused by a regression in Linux 3.0:
http://thread.gmane.org/gmane.linux.network/205326/focus=205328

https://bugzilla.gnome.org/show_bug.cgi?id=660146