obex: Fix crashes when switching VTs
We need to unregister the object when disposing it, as otherwise
we'll be trying to register the agent when there's already one
registered (to a non-existant object).

This fixes a crash when fast-user switching, or simply switching
to a VT and coming back.

See https://bugzilla.redhat.com/show_bug.cgi?id=1008653

https://bugzilla.gnome.org/show_bug.cgi?id=711628