obex: Download files through the cache dir
Save files in the cache dir before moving them to a unique name
in the Downloads folder.

We need to do this as obexd will only save to a set root folder,
which defaults to the XDG cache dir, and move this to a unique
name (without overwriting existing files).