build: Install in /usr/lib instead of /usr/lib64 on Fedora
When building gnome-user-share in Fedora, gnome-user-share-webdav.service
ends up incorrectly placed in /usr/lib64/systemd/user. It should be in
/usr/lib/systemd/user.

This is caused by us building a path from scratch, so that the prefix
could be overridden when doing an installation into a staging directory,
as the path supplied in systemd's pkgconfig file was broken in this
regard. This was however fixed in:
https://github.com/systemd/systemd/issues/9595
so use this instead.

Closes: #22